Combine some small API client types into a single file
