--- api/client/proto/duration.go
@@ -1,34 +1,0 @@
-/*
-Copyright 2020 Gravitational, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package proto
-
-import (
-	"time"
-)
-
-// Duration is a wrapper around duration
-type Duration time.Duration
-
-// Get returns time.Duration value
-func (d Duration) Get() time.Duration {
-	return time.Duration(d)
-}
-
-// Set sets time.Duration value
-func (d *Duration) Set(value time.Duration) {
-	*d = Duration(value)
-}
